chore(plugins): Change how plugin files are written
Unless necessary, like in the packer.lua file (which has not been
changed for this commit), there's no need to create a variable to get
the configuration table into and then returning that variable. Lua
supports returning the table directly, so I changed all the plugins to
do that instead.