fix(plugins): User plugins not being loaded
The way I had setup the Nebula.all_plugins table, the plugins were not
being correctly loaded. If the user added a plugin, because the index
matched, it would overwrite Nebula default plugins and prevent them for
loading. I changed the tables to go from a merge to a much simpler
table.insert and added a validation to avoid adding the same plugin
twice. The user configuration was not being loaded, so I added the
missing piece of code for that as well.