feat(config): Boostrap config loader
I want to give users the option to override plugin configs instead of
having to write a fully fledged plugin file, so I added a new helper
function to load files from the 'config' and 'user.config' namespace.
I've changed the old `load_config` and `get_config functions` to be
named `load_setup` and `get_setup`, respectively, to avoid confusion.

Migrated the configuration for the alpha plugin to this new pattern as a
POC